refactor(server): split spatial.rs into three modules to satisfy the Rust module-size ratchet
spatial.rs had grown to 511 lines, tripping module_size_ratchet (400-line
cap on non-generated .rs files) with no allowlist entry. Split along the
seams the code and spatial_tests.rs already implied:

- spatial_elevation.rs: IfcBuildingStorey.Elevation extraction with the
  ObjectPlacement Z fallback (extract_elevation_if_storey,
  extract_placement_elevation, and their attribute-index constants).
- spatial_tree.rs: the cycle/depth-guarded recursive tree walk
  (build_spatial_nodes_recursive, MAX_SPATIAL_TREE_DEPTH).
- spatial.rs: relationship-map construction (canonical_parent,
  spatial_children_map, element_containment_map) and the orphan-fill pass,
  now 315 lines.

Pure refactor: no behavior change. cargo test -p ifc-lite-server:
277 passed before and after. Verified the depth-cap orphan-fill guard
survived the split by temporarily reverting it (dropping the
canonical_parent check in the orphan-fill loop) and confirming
entities_past_the_depth_cap_are_dropped_cleanly_not_resurrected_as_fake_roots
fails, then restoring it and confirming the suite passes again.
